Display a notification when unpacking Firefox for the first time
This ensures the user does not think they need to re-launch Firefox, by
giving visual feedback that a task is being performed in the background.
We also ensure to let them know this is only for the first time, so they
are not mislead into thinking this is something that happens every time.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>