update source tarball location
the older http://ftp.mozilla.org/pub/firefox/releases/##VERSION##/source/firefox-##VERSION##.source.tar.xz
URL to define source of the releases has changed and it points to 404.
Now every release contains a SOURCE text file which points to the proper
snapshot of the source used to build the version.

This commit updates the input spec file and the update script to
follow this new process so we continue delivering the proper source for
every firefox release make use of.