don't exit if FIREFOX_INSTALL_DIR exists.
there may be cases where some preferences are added
for firefox prior to firefox being launched for the first time,
resulting in an existing FIREFOX_INSTALL_DIR but not actually
firefox setup installed.

this commit just checks if FIREFOX_INSTALL_DIR exists, and
silently creates it if it doesn't, no longer exits on existance.