firefox.sh: verify install with firefox binary
Previous to this commit, we check that the install directory
(default is ~/firefox) exists to call the firefox binary
execution, this commit changes the check to actually verify
the binary existance to assume firefox is installed.

With this change we ensure that, even if user creates the ~/firefox
directory, the binary must be in place to run firefox successfully.

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>